extracted add scripture screen into another file.
